+++ registry/datastore/testdata/fixtures/top_level_namespaces.sql
@@ -1,0 +1,3 @@
+INSERT INTO "top_level_namespaces"("id", "name", "created_at")
+VALUES (1, 'gitlab-org', '2020-03-02 17:47:39.849864+00'),
+       (2, 'a-test-group', '2020-06-08 16:01:39.476421+00');